+++ resources/views/bootstrap-4/includes/per-page.blade.php
@@ -1,11 +1,13 @@
-@if ($pagination ?? false && $showPerPage ?? false)
-    <select
-        wire:model="perPage"
-        id="perPage"
-        class="form-control"
-    >
-        <option value="10">10</option>
-        <option value="25">25</option>
-        <option value="50">50</option>
-    </select>
+@if ($paginationEnabled && $showPerPage)
+    <div class="ml-0 ml-md-3">
+        <select
+            wire:model="perPage"
+            id="perPage"
+            class="form-control"
+        >
+            <option value="10">10</option>
+            <option value="25">25</option>
+            <option value="50">50</option>
+        </select>
+    </div>
 @endif