+++ src/Traits/WithBulkActions.php
@@ -7,6 +7,8 @@ namespace Rappasoft\LaravelLivewireTables\Traits;
  */
 trait WithBulkActions
 {
+
+    public bool $showFilters = true;
     public bool $selectPage = false;
     public bool $selectAll = false;
     public $selected = [];