+++ src/Traits/WithSorting.php
@@ -9,6 +9,8 @@ use Illuminate\Database\Eloquent\Builder;
  */
 trait WithSorting
 {
+
+    public bool $showSorting = true;
     public array $sorts = [];
     public array $sortNames = [];